requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
backoff==2.2.1
certifi==2024.2.2
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.1
executing==2.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
openai==1.23.6
packaging==24.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.0.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.11.0
wcwidth==0.2.13
